Створений окремий клас для форматування дати
